feat: add jsdoc for authentication code
