+++ js/packages/teams-ai/src/authentication/OAuthAdaptiveCardAuthentication.ts
@@ -6,11 +6,26 @@ import { AdaptiveCardAuthenticationBase, AdaptiveCardLoginRequest } from './Adap
 import { AuthError, OAuthSettings } from './Authentication';
 import * as UserTokenAccess from './UserTokenAccess';
 
+/**
+ * @internal
+ * 
+ * Handles authentication for Adaptive Cards in Teams.
+ */
 export class OAuthAdaptiveCardAuthentication extends AdaptiveCardAuthenticationBase {
+
+    /**
+     * Creates a new instance of OAuthAdaptiveCardAuthentication.
+     * @param settings The OAuthSettings.
+     */
     public constructor(private readonly settings: OAuthSettings) {
         super();
     }
 
+    /**
+     * Handles the SSO token exchange.
+     * @param context The turn context.
+     * @returns A promise that resolves to the token response or undefined if token exchange failed.
+     */
     public async handleSsoTokenExchange(context: TurnContext): Promise<TokenResponse | undefined> {
         const tokenExchangeRequest = context.activity.value.authentication;
 
@@ -21,10 +36,21 @@ export class OAuthAdaptiveCardAuthentication extends AdaptiveCardAuthenticationB
         return await UserTokenAccess.exchangeToken(context, this.settings, tokenExchangeRequest);
     }
 
+    /**
+     * Handles the signin/verifyState activity.
+     * @param context The turn context.
+     * @param magicCode The magic code from sign-in.
+     * @returns A promise that resolves to undefined. The parent class will trigger silentAuth again.
+     */
     public async handleUserSignIn(context: TurnContext, magicCode: string): Promise<TokenResponse | undefined> {
         return await UserTokenAccess.getUserToken(context, this.settings, magicCode);
     }
 
+    /**
+     * Gets the sign-in link for the user.
+     * @param context The turn context.
+     * @returns A promise that resolves to the sign-in link or undefined if no sign-in link available.
+     */
     public async getLoginRequest(context: TurnContext): Promise<AdaptiveCardLoginRequest> {
         const signInResource = await UserTokenAccess.getSignInResource(context, this.settings);
         const signInLink = signInResource.signInLink;
