+++ js/packages/teams-ai/src/authentication/TeamsSsoAdaptiveCardAuthentication.ts
@@ -5,23 +5,48 @@ import { TurnContext } from "botbuilder-core";
 import { TokenResponse } from "botframework-schema";
 import { AdaptiveCardAuthenticationBase, AdaptiveCardLoginRequest } from "./AdaptiveCardAuthenticationBase";
 
+/**
+ * @internal
+ * 
+ * Handles authentication using Teams SSO for Adaptive Cards in Teams.
+ */
 export class TeamsSsoAdaptiveCardAuthentication extends AdaptiveCardAuthenticationBase {
+    /**
+     * Initializes a new instance of the TeamsSsoAdaptiveCardAuthentication class.
+     */
     public constructor() {
         super();
     }
 
+    /**
+     * Handles the SSO token exchange.
+     */
     public async handleSsoTokenExchange(): Promise<never> {
         throw new Error('Not implemented');
     }
     
+    /**
+     * Handles the user sign-in.
+     * @param context - The turn context.
+     * @param magicCode - The magic code from user sign-in.
+     */
     public async handleUserSignIn(context: TurnContext, magicCode: string): Promise<TokenResponse | undefined> {
         throw new Error('Not implemented');
     }
 
+    /**
+     * Gets the login request for Adaptive Card authentication.
+     * @param context - The turn context.
+     */
     public async getLoginRequest(context: TurnContext): Promise<AdaptiveCardLoginRequest> {
         throw new Error('Not implemented');
     }
 
+    /**
+     * Checks if the activity is valid for Adaptive Card authentication.
+     * @param context - The turn context.
+     * @returns A boolean indicating if the activity is valid.
+     */
     public override isValidActivity(context: TurnContext): boolean {
         if (super.isValidActivity(context)) {
             console.warn("TeamsSsoSetting does not support Adaptive Card authentication yet. Please use OAuthSetting instead.")
