+++ js/packages/teams-ai/src/authentication/TeamsSsoBotAuthentication.ts
@@ -6,27 +6,43 @@ import { Dialog, DialogContext, DialogTurnStatus } from "botbuilder-dialogs";
 import { Application } from "../Application";
 import { TurnState } from "../TurnState";
 import { BotAuthenticationBase } from "./BotAuthenticationBase";
-import { TeamsSsoPrompt, TeamsSsoSettings } from "./TeamsBotSsoPrompt";
+import { TeamsSsoPrompt } from "./TeamsBotSsoPrompt";
 import { DialogSet, DialogState, DialogTurnResult, WaterfallDialog } from "botbuilder-dialogs";
 import { TurnStateProperty } from "../TurnStateProperty";
 import { ConfidentialClientApplication } from "@azure/msal-node";
+import { TeamsSsoSettings } from "./TeamsSsoSettings";
 
 const SSO_DIALOG_ID = "_TeamsSsoDialog";
 
+
+/**
+ * @internal
+ * 
+ * Handles authentication for Teams bots using Single Sign-On (SSO).
+ * @template TState - The type of the turn state object.
+ */
 export class TeamsSsoBotAuthentication<TState extends TurnState> extends BotAuthenticationBase<TState> {
     private _prompt: TeamsSsoPrompt;
     private _tokenExchangeIdRegex: RegExp;
 
+    /**
+     * Initializes a new instance of the TeamsSsoBotAuthentication class.
+     * @param app - The application object.
+     * @param settings - The settings for Teams SSO.
+     * @param settingName - The name of the setting.
+     * @param msal - The MSAL (Microsoft Authentication Library) object.
+     * @param storage - The storage object for storing state.
+     */
     public constructor(
         app: Application<TState>,
-        promptSettings: TeamsSsoSettings, // Child classes will have different types for this
+        settings: TeamsSsoSettings,
         settingName: string,
         msal: ConfidentialClientApplication,
         storage?: Storage
     ) {
         super(app, settingName, storage);
 
-        this._prompt = new TeamsSsoPrompt('TeamsSsoPrompt', settingName, promptSettings, msal);
+        this._prompt = new TeamsSsoPrompt('TeamsSsoPrompt', settingName, settings, msal);
         this._tokenExchangeIdRegex = new RegExp(`[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}-${this._settingName}`);
 
         // Do not save state for duplicate token exchange events to avoid eTag conflicts
@@ -35,6 +51,13 @@ export class TeamsSsoBotAuthentication<TState extends TurnState> extends BotAuth
         });
     }
 
+    /**
+     * Run or continue the SSO dialog and returns the result.
+     * @param context - The turn context object.
+     * @param state - The turn state object.
+     * @param dialogStateProperty - The name of the dialog state property.
+     * @returns A promise that resolves to the dialog turn result containing the token response.
+     */
     public async runDialog(context: TurnContext, state: TState, dialogStateProperty: string): Promise<DialogTurnResult<TokenResponse>> {
         const dialogContext = await this.createSsoDialogContext(context, state, dialogStateProperty);
         let results = await dialogContext.continueDialog();
@@ -44,15 +67,34 @@ export class TeamsSsoBotAuthentication<TState extends TurnState> extends BotAuth
         return results;
     }
 
+    /**
+     * Continues the SSO dialog and returns the result.
+     * @param context - The turn context object.
+     * @param state - The turn state object.
+     * @param dialogStateProperty - The name of the dialog state property.
+     * @returns A promise that resolves to the dialog turn result containing the token response.
+     */
     public async continueDialog(context: TurnContext, state: TState, dialogStateProperty: string): Promise<DialogTurnResult<TokenResponse>> {
         const dialogContext = await this.createSsoDialogContext(context, state, dialogStateProperty);
         return await dialogContext.continueDialog();
     }
 
+    /**
+     * Determines whether the token exchange activity should be processed by current authentication setting.
+     * @param context - The turn context object.
+     * @returns A promise that resolves to a boolean indicating whether the token exchange route should be processed by current class instance.
+     */
     protected async tokenExchangeRouteSelector(context: TurnContext): Promise<boolean> {
         return await super.tokenExchangeRouteSelector(context) && this._tokenExchangeIdRegex.test(context.activity.value.id);
     }
 
+    /**
+     * Creates the SSO dialog context.
+     * @param context - The turn context object.
+     * @param state - The turn state object.
+     * @param dialogStateProperty - The name of the dialog state property.
+     * @returns A promise that resolves to the dialog context object.
+     */
     private async createSsoDialogContext(context: TurnContext, state: TState, dialogStateProperty: string): Promise<DialogContext> {
         const accessor = new TurnStateProperty<DialogState>(state, 'conversation', dialogStateProperty);
         const dialogSet = new DialogSet(accessor);
@@ -74,6 +116,11 @@ export class TeamsSsoBotAuthentication<TState extends TurnState> extends BotAuth
         return await dialogSet.createContext(context);
     }
 
+    /**
+     * Checks if deduplication should be performed for token exchange.
+     * @param context - The turn context object.
+     * @returns A promise that resolves to a boolean indicating whether deduplication should be performed.
+     */
     private async shouldDedup(context: TurnContext): Promise<boolean> {
         const storeItem = {
             eTag: context.activity.value.id,
@@ -93,6 +140,12 @@ export class TeamsSsoBotAuthentication<TState extends TurnState> extends BotAuth
         return false;
     }
 
+    /**
+     * Gets the storage key for storing the token exchange state.
+     * @param context - The turn context object.
+     * @returns The storage key.
+     * @throws Error if the context is invalid or the activity is not a token exchange invoke.
+     */
     private getStorageKey(context: TurnContext): string {
         if (!context || !context.activity || !context.activity.conversation) {
             throw new Error("Invalid context, can not get storage key!");
